+++ .gitignore
@@ -84,3 +84,6 @@ build-iPhoneSimulator/
 
 # unless supporting rvm < 1.11.0 or doing something fancy, ignore this:
 .rvmrc
+
+*.swo
+*.swp
